Cleanup of uneeded code

+++ get_mer_files.py
@@ -1,10 +1,14 @@
 import pylogix
 from pylogix.lgx_response import Response
-from struct import pack, unpack, unpack_from
+from struct import pack, unpack_from
 
 """
-Some other data might be returned from the first packet that has to be used
-in the "create" packet
+The goal here is to get the name of the MED file saved on a PanelView's
+\Temp\~MER.00 directory.
+
+This is accomplished by calling a DLL file that will list all files in the
+requested directory.  We'll put the results in an output file, request the
+results of the file, then delete it.
 """
 
 
@@ -37,7 +41,7 @@ def pv_test(plc, s, c, data):
 
 def create_file(plc, entry):
     """
-    Read multiple PowerFlex parameters
+    Create result file
 
     returns Response class (.TagName, .Value, .Status)
     """
@@ -68,37 +72,9 @@ def create_file(plc, entry):
     status, ret_data = plc.conn.send(request, False)
     return Response(None, None, status)
 
-def seems_important(plc):
-    """
-    Not sure what this does yet, but if I know comments, this
-    will exist forever
-    """
-    conn = plc.conn.connect(False)
-    if not conn[0]:
-        return Response(None, None, conn[1])
-
-    cip_service = 0x53
-    cip_service_size = 0x03
-    cip_class_type = 0x21
-    cip_class = 0x04ff
-    cip_instance_type = 0x24
-    cip_instance = 0x01
-
-    request = pack('<BBHHBBI',
-                    cip_service,
-                    cip_service_size,
-                    cip_class_type,
-                    cip_class,
-                    cip_instance_type,
-                    cip_instance,
-                    0x02)
-
-    status, ret_data = plc.conn.send(request, False)
-    return Response(None, None, status)
-
 def delete_file(plc):
     """
-    Read multiple PowerFlex parameters
+    Delete result file
 
     returns Response class (.TagName, .Value, .Status)
     """
@@ -124,9 +100,9 @@ def delete_file(plc):
     status, ret_data = plc.conn.send(request, False)
     return Response(None, None, status)
 
-def get_mers(plc):
+def get_med(plc):
     """
-    Read multiple PowerFlex parameters
+    Request the results of the file that was generated
 
     returns Response class (.TagName, .Value, .Status)
     """
@@ -165,35 +141,24 @@ def get_mers(plc):
 
     return Response(None, file_name, status)
 
-with pylogix.PLC("192.168.1.12") as comm:
 
-    pc = "HKEY_LOCAL_MACHINE\\Software\\Rockwell Software\\RSLinxNG\\CIP Identity\\ProductCode\0"
-    pt = "HKEY_LOCAL_MACHINE\\SOFTWARE\\Rockwell Software\\RSLinxNG\\CIP Identity\\ProductType\0"
+with pylogix.PLC("192.168.1.12") as comm:
 
     init_file = "\\Windows\\RemoteHelper.DLL\0FileBrowse\0\\Temp\\~MER.00\\*.med::\\Application Data\\Rockwell Software\\RSViewME\\Runtime\\DillyDilly.txt\0"
     file_location = "\\Application Data\\Rockwell Software\\RSViewME\\Runtime\\DillyDilly.txt\0"
     uninit_file = "\\Windows\\RemoteHelper.DLL\0DeleteRemFile\0\\Application Data\\Rockwell Software\\RSViewME\\Runtime\\DillyDilly.txt\0"
 
     response = pv_test(comm, 0x50, 0x04fd, init_file)
-    print(response)
+    if response.Status == "Success":
+        response = create_file(comm, file_location)
 
-    response = create_file(comm, file_location)
-    print(response)
+    if response.Status == "Success":
+        response = get_med(comm)
+        print(response)
 
-    response = pv_test(comm, 0x51, 0x04fe, pt)
-    print(response)
+    if response.Status == "Success":
+        response = delete_file(comm)
+    
+    if response.Status == "Success":
+        response = pv_test(comm, 0x50, 0x04fd, uninit_file)
 
-    response = pv_test(comm, 0x51, 0x04fe, pc)
-    print(response)
-
-    response = get_mers(comm)
-    print(response)
-
-    response = seems_important(comm)
-    print(response)
-
-    response = delete_file(comm)
-    print(response)
-
-    response = pv_test(comm, 0x50, 0x04fd, uninit_file)
-    print(response)
